# Rid pointless wrapping with BufferedInputStream from class reading